Extract SWRChart computation logic into swrChartUtils.ts
This refactoring removes inline `useMemo` computation blocks from `SWRChart.tsx`
and extracts them into pure, standalone functions in a new file `swrChartUtils.ts`.
This separation of concerns makes `SWRChart.tsx` much shorter and primarily
focused on React rendering and lifecycle, improving readability and maintainability
without altering the functional behavior.
